Seed the database and change price property to double

+++ Models/Check.cs
@@ -3,7 +3,7 @@
     public class Check : DbObject
     {
         public List<Item> Items { get; set; }
-        public int Sum { get; set; }
+        public double Sum { get; set; }
         public string Location { get; set; }
         public string? VerificationPath { get; set; }
     }
